requirements.txt
pyTelegramBotAPI
websocket-client
